Extract outline query runner partial
