+++ src/CodeIndex/Cli/QueryCommandRunner.cs
@@ -5607,228 +5607,6 @@ public static partial class QueryCommandRunner
         return "No definition body content is available for the matched definitions.";
     }
 
-    public static int RunOutline(string[] cmdArgs, JsonSerializerOptions jsonOptions)
-    {
-        if (cmdArgs.Length == 0 || cmdArgs[0].StartsWith('-'))
-        {
-            WriteUsageError(
-                "outline requires a file path.",
-                GetUsageLineOrThrow("outline"),
-                "Pass the indexed file path, for example: `cdidx outline src/CodeIndex/Program.cs`.");
-            return CommandExitCodes.UsageError;
-        }
-
-        var previewOptionError = ValidatePreviewOptions("outline", cmdArgs[1..], allowMaxLineWidth: false, allowFocusOptions: false);
-        if (previewOptionError != null)
-        {
-            CommandErrorWriter.WriteStderr(previewOptionError);
-            return CommandExitCodes.UsageError;
-        }
-        var options = ParseArgs(
-            cmdArgs[1..],
-            jsonDefault: false,
-            validateDefaultLimit: false,
-            validateDefaultSnippetLines: false,
-            validateDefaultMaxLineWidth: false,
-            allowOutlineSort: true);
-        if (TryWriteUnsupportedOptionError("outline", cmdArgs[1..], CliFlagSchema.GetAcceptedFlagNamesForCommand("outline")))
-            return CommandExitCodes.UsageError;
-        if (TryWriteParseError(options, "outline"))
-            return CommandExitCodes.UsageError;
-        if (TryWriteInvalidOutlineKindFilterError(options))
-            return CommandExitCodes.InvalidArgument;
-        if (TryWriteUnexpectedPositionals("outline", options))
-            return CommandExitCodes.UsageError;
-        if (options.SearchCursor.HasValue || options.UnusedCursorOffset.HasValue)
-        {
-            WriteUsageError(
-                "outline --cursor must use an outline pagination cursor.",
-                GetUsageLineOrThrow("outline"),
-                "Use the `next_cursor` value returned by `cdidx outline <path> --json --limit <n>`.");
-            return CommandExitCodes.UsageError;
-        }
-
-        var filePath = DbPathResolver.ResolveQueryFilePath(options.DbPath, cmdArgs[0], options.DbPathExplicit);
-        var outlineSortMode = options.SortExplicit && options.SortValue != null && TryParseOutlineSortMode(options.SortValue, out var parsedSortMode)
-            ? parsedSortMode
-            : OutlineSortMode.Source;
-        var includeReferenceCounts = OutlineNeedsReferenceCounts(options, outlineSortMode);
-        var includeDerivedMetadata = OutlineNeedsDerivedMetadata(options, outlineSortMode);
-        return WithDb(options, jsonOptions, reader =>
-        {
-            var outline = reader.GetOutline(filePath, includeReferenceCounts: includeReferenceCounts);
-            if (outline == null)
-            {
-                if (options.Json)
-                    Console.WriteLine(JsonSerializer.Serialize(new QueryPathErrorJsonResult(filePath, "file not found in index"), CliJsonSerializerContextFactory.Create(jsonOptions).QueryPathErrorJsonResult));
-                else
-                    CommandErrorWriter.WriteStderr($"Error: '{filePath}' not found in index.");
-                return CommandExitCodes.NotFound;
-            }
-
-            var kindFilters = BuildOutlineKindFilters(options.Kind);
-            var filteredSymbols = ApplyOutlineKindFilters(outline.Symbols, kindFilters);
-            var displaySourceSymbols = ApplyOutlineSort(filteredSymbols, outlineSortMode, includeDerivedMetadata);
-            if (options.Json)
-            {
-                if (options.Compact)
-                {
-                    var payload = BuildOutlineJsonPayload(outline, displaySourceSymbols, kindFilters, outlineSortMode, options, jsonOptions, compact: true);
-                    Console.WriteLine(payload.ToJsonString(jsonOptions));
-                }
-                else if (HasOutlineJsonControls(options, kindFilters))
-                {
-                    var payload = BuildOutlineJsonPayload(outline, displaySourceSymbols, kindFilters, outlineSortMode, options, jsonOptions, compact: false);
-                    Console.WriteLine(payload.ToJsonString(jsonOptions));
-                }
-                else
-                {
-                    var payload = JsonSerializer.SerializeToNode(outline, CliJsonSerializerContextFactory.Create(jsonOptions).OutlineResult)!.AsObject();
-                    Console.WriteLine(payload.ToJsonString(jsonOptions));
-                }
-            }
-            else
-            {
-                var outlineContent = reader.GetExcerpt(filePath, 1, outline.TotalLines)?.Content;
-
-                Console.WriteLine($"# {outline.Path}  ({outline.Lang ?? "unknown"}, {outline.TotalLines} lines, {filteredSymbols.Count} symbols)");
-                Console.WriteLine();
-                var duplicateNames = filteredSymbols
-                    .GroupBy(sym => sym.Name, StringComparer.Ordinal)
-                    .Where(group => group.Count() > 1)
-                    .Select(group => group.Key)
-                    .ToHashSet(StringComparer.Ordinal);
-                var displaySymbols = ApplyOutlineHumanPaging(displaySourceSymbols, options);
-                foreach (var sym in displaySymbols)
-                {
-                    // Indent nested symbols by computed tree depth / コンテナ連鎖の深さでインデント
-                    var indent = sym.Depth > 0 ? new string(' ', 4 * sym.Depth) : "";
-                    var useDisplayName = sym.Kind is "function" or "method" or "constructor"
-                        && duplicateNames.Contains(sym.Name)
-                        && !string.IsNullOrWhiteSpace(sym.DisplayName);
-                    var ret = !useDisplayName && sym.ReturnType != null ? $": {sym.ReturnType} " : "";
-                    var sig = useDisplayName ? sym.DisplayName : sym.Signature ?? $"{sym.Kind} {sym.Name}";
-                    // Avoid duplicating visibility when signature already contains it
-                    // シグネチャに既に visibility が含まれている場合は重複を避ける
-                    var vis = !useDisplayName && sym.Visibility != null && !sig.TrimStart().StartsWith(sym.Visibility, StringComparison.Ordinal)
-                        ? $"{sym.Visibility} "
-                        : "";
-                    Console.WriteLine($"  {sym.Line,5}  {indent}{vis}{sig} {ret}");
-                }
-
-                // AI-orientation hint for C# files that look like top-level-statements programs:
-                // no class / struct / interface / enum / namespace / record / delegate at all
-                // means the executable body lives between the imports and local functions and
-                // will not appear in outline at all. Emitting a short note on stderr keeps the
-                // main human-readable block clean while giving AI consumers a reason for the gap.
-                // AI向けヒント: C# のトップレベルステートメント想定のファイル
-                // （class / struct / interface / enum / namespace / record / delegate が一切無い）は、
-                // 実行本体が import と local function の間に書かれるため outline に現れない。
-                // 人間向け本体を汚さないよう、理由を短く stderr に出す。
-                if (LooksLikeCsharpTopLevelStatements(outline, outlineContent))
-                {
-                    CommandErrorWriter.WriteStderr();
-                    CommandErrorWriter.WriteStderr("Note: no type/namespace declarations found; this file likely uses C# top-level statements.");
-                    CommandErrorWriter.WriteStderr("      Outline lists imports and local functions only; the executable body is not indexed as symbols.");
-                }
-            }
-            return CommandExitCodes.Success;
-        });
-    }
-
-    /// <summary>
-    /// Heuristic: hint only when a non-trivial C# file has no type/namespace declarations and
-    /// its reconstructed content still contains uncovered file-scope executable code after
-    /// skipping symbol-covered lines, imports, metadata-only attribute lines, comments, and
-    /// preprocessor directives. This keeps the note off common files such as GlobalUsings.cs,
-    /// AssemblyInfo.cs, and local-function-only files while preserving statement-only Program.cs
-    /// files.
-    /// Tiny files (snippets, partials under ~20 lines) are excluded to avoid noise.
-    /// ヒューリスティック: 20 行以上の C# ファイルで型/名前空間宣言が無く、かつ
-    /// import 行、metadata-only 属性行、コメント、プリプロセッサ行を除いても
-    /// file-scope の実行コードが残る場合だけヒントを出す。これにより GlobalUsings.cs や
-    /// AssemblyInfo.cs の誤検出を避けつつ、
-    /// statement-only の Program.cs は拾い続ける。小さい断片はノイズ回避のため除外。
-    /// </summary>
-    private static bool LooksLikeCsharpTopLevelStatements(OutlineResult outline, string? content)
-    {
-        if (outline.Lang != "csharp") return false;
-        if (outline.TotalLines < 20) return false;
-        foreach (var sym in outline.Symbols)
-        {
-            if (sym.Kind is "class" or "struct" or "interface" or "enum" or "namespace" or "delegate" or "record")
-                return false;
-        }
-
-        if (string.IsNullOrWhiteSpace(content))
-            return false;
-
-        var coveredLines = new bool[Math.Max(outline.TotalLines, 0) + 1];
-        foreach (var sym in outline.Symbols)
-        {
-            var startLine = sym.StartLine > 0 ? sym.StartLine : sym.Line;
-            var endLine = sym.EndLine >= startLine ? sym.EndLine : startLine;
-            startLine = Math.Max(1, startLine);
-            endLine = Math.Min(outline.TotalLines, endLine);
-            for (var lineNumber = startLine; lineNumber <= endLine; lineNumber++)
-                coveredLines[lineNumber] = true;
-        }
-
-        var inBlockComment = false;
-        var currentLineNumber = 0;
-        foreach (var rawLine in content.Split('\n'))
-        {
-            currentLineNumber++;
-            var line = rawLine.Trim();
-            if (line.Length == 0)
-                continue;
-            if (currentLineNumber < coveredLines.Length && coveredLines[currentLineNumber])
-                continue;
-
-            if (inBlockComment)
-            {
-                if (line.Contains("*/", StringComparison.Ordinal))
-                    inBlockComment = false;
-                continue;
-            }
-
-            if (line.StartsWith("/*", StringComparison.Ordinal))
-            {
-                if (!line.Contains("*/", StringComparison.Ordinal))
-                    inBlockComment = true;
-                continue;
-            }
-
-            if (line.StartsWith("using ", StringComparison.Ordinal))
-            {
-                if (line.StartsWith("using var ", StringComparison.Ordinal))
-                    return true;
-                if (line.StartsWith("using (", StringComparison.Ordinal))
-                    return true;
-                continue;
-            }
-            if (line.StartsWith("global using ", StringComparison.Ordinal))
-                continue;
-            if (line.StartsWith("extern alias ", StringComparison.Ordinal))
-                continue;
-            if (line.StartsWith("[assembly:", StringComparison.Ordinal))
-                continue;
-            if (line.StartsWith("[module:", StringComparison.Ordinal))
-                continue;
-            if (line.StartsWith("//", StringComparison.Ordinal))
-                continue;
-            if (line.StartsWith("*", StringComparison.Ordinal))
-                continue;
-            if (line.StartsWith("*/", StringComparison.Ordinal))
-                continue;
-            if (line.StartsWith("#", StringComparison.Ordinal))
-                continue;
-            return true;
-        }
-
-        return false;
-    }
-
     public static int RunStatus(
         string[] cmdArgs,
         JsonSerializerOptions jsonOptions,
